requirements.txt
httpx==0.24.1
python-dotenv==1.0.0
google-generativeai
fastapi
pillow
aiohttp
asyncio
uvicorn
